overlay: Fix automatic overlay closing on click when <select> is active.
When `<select>` dropdown is open, click events outside `<select>` is
registered to common parent node of <select> and element on which
user clicked. This leads to overlay exiting sometimes when click
was registered on .overlay-content.

This commit fixes this behaviour by making settings overlay html
structure match that of channel settings overlay and remove
overlay-content used for transition.

Fixes: zulip#33464.